feat(organizer): add policy to allow only owner to update settings

+++ app/Http/Controllers/OrganizerController.php
@@ -7,6 +7,7 @@ use App\Models\Organizer;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class OrganizerController extends Controller
 {
@@ -60,10 +61,16 @@ class OrganizerController extends Controller
     }
 
     public function edit(Organizer $organizer) {
+
+        Gate::authorize('viewSettings', $organizer);
+
         return view('organizer.settings', compact('organizer'));
     }
 
     public function update(Request $request, Organizer $organizer) {
+        
+        Gate::authorize('update', $organizer);
+
         $request->validate([
             'name' => ['required', 'string', 'min:1', 'max:255'],
             'profile' => ['nullable', 'image', 'mimes:jpeg,png,jpg,gif,svg','max:2048'],
@@ -88,6 +95,9 @@ class OrganizerController extends Controller
     }
 
     public function destroy(Organizer $organizer) {
+        
+        Gate::authorize('delete', $organizer);
+        
         $organizer->delete();
 
         return redirect()->route('organizer.home');
